requirements.txt
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.5
Flask==3.0.0
Flask-Cors==4.0.0
GeoAlchemy2==0.14.3
geopandas==0.14.1
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
psycopg2==2.9.9
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
